🚀 feat(cron.go): Add new cron rule for bookmark search and update types/search.go 🐛 fix(search.go): Change CreatedAt to int64 in types/search.go 🐛 fix(command.go): Update CreatedAt to Unix timestamp in command.go

+++ internal/bots/dev/command.go
@@ -210,7 +210,7 @@ var commandRules = []command.Rule{
 				Title:       "test....",
 				Description: "the....",
 				Url:         "/url/test",
-				CreatedAt:   int32(time.Now().Unix()),
+				CreatedAt:   time.Now().Unix(),
 			})
 			if err != nil {
 				return types.TextMsg{Text: err.Error()}
